Fold LimbSolver into Limb
