+++ src/main/java/com/marginallyclever/ro3/node/nodes/behavior/actions/LimbMoveToTargetPanel.java
@@ -2,7 +2,6 @@ package com.marginallyclever.ro3.node.nodes.behavior.actions;
 
 import com.marginallyclever.ro3.PanelHelper;
 import com.marginallyclever.ro3.apps.nodeselector.NodeSelector;
-import com.marginallyclever.ro3.node.nodes.limbsolver.LimbSolver;
 import com.marginallyclever.ro3.node.nodes.pose.Pose;
 
 import javax.swing.*;
@@ -24,15 +23,6 @@ public class LimbMoveToTargetPanel extends JPanel {
                 limbMoveToTarget.setTarget(selector.getSubject());
             });
             PanelHelper.addLabelAndComponent(this, "Target", selector);
-        }
-
-        // solver
-        {
-            NodeSelector<LimbSolver> selector = new NodeSelector<>(LimbSolver.class, limbMoveToTarget.getSolver());
-            selector.addPropertyChangeListener("subject", (evt) -> {
-                limbMoveToTarget.setSolver(selector.getSubject());
-            });
-            PanelHelper.addLabelAndComponent(this, "LimbSolver", selector);
         }
     }
 }
