+++ src/test/java/com/marginallyclever/ro3/node/nodes/marlinrobot/marlinrobotarm/MarlinRobotArmTest.java
@@ -2,7 +2,6 @@ package com.marginallyclever.ro3.node.nodes.marlinrobot.marlinrobotarm;
 
 import com.marginallyclever.convenience.helpers.StringHelper;
 import com.marginallyclever.ro3.Registry;
-import com.marginallyclever.ro3.node.nodes.limbsolver.LimbSolver;
 import com.marginallyclever.ro3.node.nodes.marlinrobot.MarlinListener;
 import com.marginallyclever.ro3.node.nodes.pose.Pose;
 import com.marginallyclever.ro3.node.nodes.pose.poses.Limb;
@@ -17,14 +16,12 @@ class MarlinRobotArmTest {
     @Test
     void testGetTarget() {
         Limb limb = new Limb();
-        LimbSolver limbSolver = new LimbSolver();
         MarlinRobotArm marlinRobotArm = new MarlinRobotArm();
         Pose target = new Pose();
         limb.addChild(marlinRobotArm);
-        limb.addChild(limbSolver);
         limb.addChild(target);
-        limbSolver.setTarget(target);
-        marlinRobotArm.setSolver(limbSolver);
+        limb.setTarget(target);
+        marlinRobotArm.setLimb(limb);
         assertEquals(target, marlinRobotArm.getTarget());
     }
 
@@ -43,13 +40,9 @@ class MarlinRobotArmTest {
     void testSetLimbAndSolver() {
         Limb limb = new Limb();
         MarlinRobotArm marlinRobotArm = new MarlinRobotArm();
-        LimbSolver limbSolver = new LimbSolver();
         limb.addChild(marlinRobotArm);
-        limb.addChild(limbSolver);
         marlinRobotArm.setLimb(limb);
-        marlinRobotArm.setSolver(limbSolver);
         assertEquals(limb, marlinRobotArm.getLimb().getSubject());
-        assertEquals(limbSolver, marlinRobotArm.getSolver().getSubject());
     }
 
     @Test
@@ -57,20 +50,16 @@ class MarlinRobotArmTest {
         Registry.start();
 
         var limb = new Limb("a");
-        var solver = new LimbSolver("b");
 
         MarlinRobotArm mra = new MarlinRobotArm();
         mra.addChild(limb);
-        mra.addChild(solver);
         mra.setLimb(limb);
-        mra.setSolver(solver);
 
         var json = mra.toJSON();
         MarlinRobotArm mra2 = new MarlinRobotArm();
         mra2.fromJSON(json);
 
         assertEquals(mra.getLimb().getSubject().getName(),mra2.getLimb().getSubject().getName());
-        assertEquals(mra.getSolver().getSubject().getName(),mra2.getSolver().getSubject().getName());
     }
 
     @Test
